new table in DB. Review exist models and it methods.

+++ data/order.go
@@ -1,10 +1,12 @@
 package data
 
 import (
+	"errors"
 	"log"
 	"time"
 )
 
+//Order struct for "orders" table
 type Order struct {
 	ID      int
 	Title   string
@@ -12,6 +14,22 @@ type Order struct {
 	Customer
 	Status
 	CreatedAt time.Time
+}
+
+//PerformedOrder struct for "performed_order" table
+type PerformedOrder struct {
+	ID int
+}
+
+//CompleteOrder struct for "complete_order" table
+type CompleteOrder struct {
+	ID int
+}
+
+//Status struct for "order_status" table
+type Status struct {
+	ID   int
+	Name string
 }
 
 //CreateOrder new row in "order" table
@@ -30,18 +48,23 @@ func (customer *Customer) CreateOrder(title string, content string) (order Order
 }
 
 //UpdateOrder row in "order" table
-func (customer *Customer) UpdateOrder(order Order) (err error) {
+func (customer *Customer) UpdateOrder(order *Order) (err error) {
 	if order.Customer.ID == customer.ID {
-		statement := `UPDATE orders SET title = $1, content = $2 WHERE id = $3`
-		stmt, err := Db.Prepare(statement)
-		if err != nil {
-			panic(err)
-		}
+		if order.IsAvailable() {
+			statement := `UPDATE orders SET title = $1, content = $2 WHERE id = $3
+			returning id`
+			stmt, err := Db.Prepare(statement)
+			if err != nil {
+				return err
+			}
 
-		defer stmt.Close()
-		return stmt.QueryRow(order.Title, order.Content).Scan(&order.ID)
+			defer stmt.Close()
+			stmt.QueryRow(order.Title, order.Content, order.ID).Scan(&order.ID)
+		} else {
+			return errors.New("Can't change order, it status not 'Available'")
+		}
 	} else {
-		log.Println("Insufficient rights to delete order")
+		return errors.New("Insufficient rights to update order")
 	}
 	return
 }
@@ -49,16 +72,67 @@ func (customer *Customer) UpdateOrder(order Order) (err error) {
 // DeleteOrder row from "order" table
 func (customer *Customer) DeleteOrder(order Order) (err error) {
 	if order.Customer.ID == customer.ID {
-		statement := "DELETE FROM orders WHERE id = $1"
+		if order.IsAvailable() {
+			statement := "DELETE FROM orders WHERE id = $1"
+			stmt, err := Db.Prepare(statement)
+			if err != nil {
+				return err
+			}
+
+			defer stmt.Close()
+			_, err = stmt.Exec(order.ID)
+			if err != nil {
+				return err
+			}
+		} else {
+			return errors.New("Can't change order, it status not 'Available'")
+		}
+	} else {
+		return errors.New("Insufficient rights to delete order")
+	}
+	return
+}
+
+//MakeOrderPerformed row in "order" table
+func (customer *Customer) MakeOrderPerformed(order *Order) (err error) {
+	if order.Customer.ID == customer.ID {
+		statement := `UPDATE orders SET status_id = 2 WHERE id = $1
+									returning id`
 		stmt, err := Db.Prepare(statement)
 		if err != nil {
 			log.Println(err)
 		}
 
 		defer stmt.Close()
-		_, err = stmt.Exec(order.ID)
+		stmt.QueryRow(order.ID).Scan(&order.ID)
 	} else {
-		log.Println("Insufficient rights to delete order")
+		return errors.New("Insufficient rights make order performed")
+	}
+	return
+}
+
+//MakeOrderDone row in "order" table
+func (customer *Customer) MakeOrderDone(order *Order) (err error) {
+	if order.Customer.ID == customer.ID {
+		statement := `UPDATE orders SET status_id = 2 WHERE id = $1
+									returning id`
+		stmt, err := Db.Prepare(statement)
+		if err != nil {
+			log.Println(err)
+		}
+
+		defer stmt.Close()
+		stmt.QueryRow(order.ID).Scan(&order.ID)
+	} else {
+		return errors.New("Insufficient rights make order done")
+	}
+	return
+}
+
+//IsAvailable - check order status, if "available" return true
+func (order *Order) IsAvailable() (available bool) {
+	if order.Status.Name == "Available" {
+		available = true
 	}
 	return
 }
